ecom/requirements.txt
asgiref==3.2.10
Babel==2.8.0
dj-database-url==0.5.0
Django==3.0.10
django-crispy-forms==1.9.2
django-debug-toolbar==3.1
django-extra-views==0.13.0
django-haystack==3.0
django-heroku==0.3.1
django-oscar==2.1
django-oscar-api==2.0.2
django-phone-field==1.8.1
django-phonenumber-field==3.0.1
django-rest-framework==0.1.0
django-tables2==2.2.1
django-treebeard==4.3.1
django-widget-tweaks==1.4.8
djangorestframework==3.11.1
factory-boy==2.12.0
Faker==4.1.3
gunicorn==20.0.4
phonenumbers==8.12.9
Pillow==7.2.0
psycopg2==2.8.6
purl==1.5
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sorl-thumbnail==12.6.3
sqlparse==0.3.1
text-unidecode==1.3
whitenoise==5.2.0
